backend/requirements.txt
anyio==3.6.2
astroid==2.15.5
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0       
click==8.1.3
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
dnspython==2.3.0
executing==1.2.0
fastapi==0.95.1
filelock==3.12.0
fsspec==2023.5.0
google-api-core==2.11.0
google-auth==2.19.0
google-cloud-core==2.3.2        
google-cloud-storage==2.9.0     
google-crc32c==1.5.0
google-resumable-media==2.5.0   
googleapis-common-protos==1.59.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.14.1
idna==3.4
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.13.2
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0     
lazy-object-proxy==1.9.0
loguru==0.7.0
MarkupSafe==2.1.2       
matplotlib-inline==0.1.6
mccabe==0.7.0
mpmath==1.3.0
nest-asyncio==1.5.6     
networkx==3.1
nltk==3.8.1
numpy==1.24.3
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
Pygments==2.15.1
pylint==2.17.4
PyPDF2==3.0.1
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
regex==2023.5.5
requests==2.30.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
sympy==1.12
threadpoolctl==3.1.0
tokenizers==0.13.3
tomlkit==0.11.8
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
typing_extensions==4.5.0
urllib3==1.26.16
uvicorn==0.22.0
wcwidth==0.2.6
win32-setctime==1.1.0
wrapt==1.15.0